libmv_core tests in progress (an other commit will occurs in few hours)

+++ trunk/otherProjects/libmv_core/tests/fundamental_test.cc
@@ -101,7 +101,7 @@ TEST(Fundamental, MotionFromEssential) {
   vector<Vec3> ts;
   MotionFromEssential(E, &Rs, &ts);
   bool one_solution_is_correct = false;
-  for (size_t i = 0; i < Rs.size(); ++i) {
+  for (int i = 0; i < Rs.size(); ++i) {
     if (FrobeniusDistance(Rs[i], R) < 1e-8 && DistanceL2(ts[i], t) < 1e-8) {
       one_solution_is_correct = true;
       break;
